third_party/WebKit/LayoutTests/http/tests/security/suborigins/suborigin-valid-names.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>Valid suborigin names</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<body>
<script>
var test_suborigin_names = [
  'foobar',
  'foob4r',
];

async_test(t => {
    var i = 0;

    window.onmessage = t.step_func(function(event) {
        assert_equals(
          event.data,
          'SecurityError: Blocked a frame with origin ' +
          '"http://127.0.0.1:8000" from accessing a cross-origin frame.');
        next();
      });

    function next() {
      if (i >= test_suborigin_names.length) {
        t.done();
        return;
      }

      var iframe = document.createElement('iframe');
      iframe.src =
        'resources/reach-into-iframe.php?childsuborigin=' +
        test_suborigin_names[i];
      document.body.appendChild(iframe);

      i++;
    }

    next();
  }, 'Valid suborigin names');
</script>
</body>
